Write snapshots to disk

+++ src/main/scala/net/bstjohn/ad/generator/format/domains/DomainProperties.scala
@@ -10,7 +10,7 @@ case class DomainProperties(
   domainsid: String,
   highvalue: Boolean,
   description: Option[String],
-  whencreated: Int,
+  whencreated: Long,
   functionallevel: String,
 )
