+++ src/main/scala/net/bstjohn/ad/generator/format/groups/GroupProperties.scala
@@ -10,7 +10,7 @@ case class GroupProperties(
   domainsid: String,
   highvalue: Option[Boolean],
   description: Option[String],
-  whencreated: Option[Int],
+  whencreated: Option[Long],
   admincount: Option[Boolean],
 )
